+++ src/piplexed/display/print_table.py
@@ -6,13 +6,13 @@ from rich import print as rich_print
 from rich.table import Table
 from rich.text import Text
 
-from piplexed.pipx_venvs import PackageInfo
+from piplexed.venvs import PackageInfo
 
 
-def print_list_table(packages: Iterable[PackageInfo]) -> None:
-    table = Table(title="Pip❎ Packages")
+def print_installed_table(packages: Iterable[PackageInfo], tool: str) -> None:
+    table = Table(title=f"{tool.upper()} Packages")
     table.add_column("Package Name", justify="right", style="dark_orange", no_wrap=True)
-    table.add_column("Pipx Version", justify="right", style="deep_sky_blue1", no_wrap=True)
+    table.add_column("Installed Version", justify="right", style="deep_sky_blue1", no_wrap=True)
     table.add_column("Python Version", justify="right", style="green4", no_wrap=True)
 
     for pkg in packages:
@@ -21,10 +21,10 @@ def print_list_table(packages: Iterable[PackageInfo]) -> None:
     rich_print(table)
 
 
-def print_outdated_table(package_data: Iterable[PackageInfo]) -> None:
-    table = Table(title="Pip❌ Outdated Packages")
+def print_outdated_table(package_data: Iterable[PackageInfo], tool: str) -> None:
+    table = Table(title=f"{tool.upper()} Outdated Packages")
     table.add_column("Package Name", justify="right", style="dark_orange", no_wrap=True)
-    table.add_column("Pipx Version", justify="right", style="deep_sky_blue1", no_wrap=True)
+    table.add_column("Installed Version", justify="right", style="deep_sky_blue1", no_wrap=True)
     table.add_column("PyPI Version", justify="right", style="red3", no_wrap=True)
 
     for pkg in package_data: