+++ src/piplexed/pypi/pypi_info.py
@@ -14,16 +14,17 @@ from rich.progress import TaskProgressColumn
 from rich.progress import TextColumn
 from rich.progress import TimeRemainingColumn
 
-from piplexed.pipx_venvs import PackageInfo
-from piplexed.pipx_venvs import get_pipx_metadata
+from piplexed.venvs import PackageInfo
 
 
 def pypi_package_info(client: PyPISimple, package_name: str) -> list[DistributionPackage]:
+    """List of installable packages from PyPI for a given project"""
     project_page = client.get_project_page(package_name)
     return project_page.packages
 
 
-def latest_pypi_version(pkg_versions: list[DistributionPackage], stable: bool) -> Version:  # noqa: FBT001
+def latest_pypi_version(pkg_versions: list[DistributionPackage], is_prerelease: bool) -> Version:  # noqa: FBT001
+    """Finds most recent package version that has an sdist distribution"""
     versions: list[Version] = []
     for pkg in pkg_versions:
         if pkg.version is not None and pkg.package_type == "sdist" and not pkg.is_yanked:
@@ -31,7 +32,7 @@ def latest_pypi_version(pkg_versions: list[DistributionPackage], stable: bool) -
                 pkg_vsn = Version(pkg.version)
             except InvalidVersion:
                 continue
-            if not pkg_vsn.is_prerelease or not stable:
+            if not pkg_vsn.is_prerelease or is_prerelease:
                 versions.append(pkg_vsn)
 
     # use max(Version) instead of reversing order of package_page as recommended in PEP 700
@@ -39,19 +40,19 @@ def latest_pypi_version(pkg_versions: list[DistributionPackage], stable: bool) -
     return max(versions)
 
 
-def get_pypi_versions(client: PyPISimple, package: PackageInfo, stable: bool) -> PackageInfo:  # noqa: FBT001
+def get_pypi_versions(client: PyPISimple, package: PackageInfo, is_prerelease: bool) -> PackageInfo:  # noqa: FBT001
+    """Update a PackageInfo dataclass with the most recent version available on PyPI"""
     pypi_versions: list[DistributionPackage] = pypi_package_info(client=client, package_name=package.name)
 
-    latest_version = latest_pypi_version(pypi_versions, stable=stable)
+    latest_version = latest_pypi_version(pypi_versions, is_prerelease=is_prerelease)
 
     package.latest_pypi_version = latest_version
 
     return package
 
 
-def find_outdated_packages(*, stable: bool = True) -> list[PackageInfo]:
-    venvs: list[PackageInfo] = get_pipx_metadata()
-
+def find_most_recent_version_on_pypi(*, venvs: list[PackageInfo], is_prerelease: bool) -> list[PackageInfo]:
+    """Get most recent version on PyPI for a given list of packages (PackageInfo dataclasses)"""
     with ExitStack() as stack:
         client = stack.enter_context(PyPISimple())
         progress_bar = Progress(
@@ -67,7 +68,7 @@ def find_outdated_packages(*, stable: bool = True) -> list[PackageInfo]:
 
         executor = stack.enter_context(ThreadPoolExecutor(max_workers=5))
 
-        results = [executor.submit(get_pypi_versions, client, pkg, stable) for pkg in venvs]
+        results = [executor.submit(get_pypi_versions, client, pkg, is_prerelease) for pkg in venvs]
 
         updates = []
         for future in as_completed(results):